+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fixed: Edit permissions checks for Related to Show post assignments
 * Fixed: Main Language option value for WordPress Setting
 * Fixed: make Date on Tab clickable on Tabbed Schedule View
+* Fixed: prevent possible conflicts with changes not saved reload message
+* Fixed: do not conflict check Shift against self for last shift check
 
 ### 2.3.3.5
 * Fixed: use schedule based on start_day if specified for Schedule view